Refactoring.. better naming of classes
